refactor: move stuff to presentation
because it makes more sense to put them here, I think
